feat: install and setup babel and move main files to src
