+++ .gitignore
@@ -4,3 +4,4 @@ node_modules/
 package-lock.json
 src/commands/Dev/MoonCommand.js
 logs/
+dist/
