+++ package.json
@@ -2,10 +2,11 @@
   "name": "menhera-bot",
   "version": "2.8.3-stable",
   "description": "Uma bot brasileira de RPG para o Discord",
-  "main": "shard.js",
+  "main": "src/shard.ts",
   "scripts": {
-    "start": "ts-node shard.ts",
-    "dev": "nodemon shard.js",
+    "build": "rm -rf dist/* && babel src --extensions \".js,.ts\" --ignore *.spec.js --out-dir dist --copy-files --source-maps --verbose",
+    "start": "node dist/shard.js",
+    "dev": "ts-node-dev --respawn --transpile-only --ignore-watch node_modules --no-notify src/shard.ts",
     "lint:fix": "eslint --fix .",
     "test": "jest"
   },
@@ -38,11 +39,16 @@
   "license": "AGPL-3.0",
   "readme": "https://github.com/ySnoopyDogy/MenheraBot#readme",
   "devDependencies": {
+    "@babel/cli": "^7.14.3",
+    "@babel/core": "^7.14.3",
+    "@babel/preset-env": "^7.14.2",
+    "@babel/preset-typescript": "^7.13.0",
     "eslint": "^7.13.0",
     "eslint-config-airbnb-base": "^14.2.1",
     "eslint-plugin-import": "^2.22.1",
     "jest": "^26.6.3",
     "ts-node": "^9.1.1",
+    "ts-node-dev": "^1.1.6",
     "typescript": "^4.2.4"
   }
 }
