+++ tsconfig.json
@@ -5,6 +5,11 @@
     "target": "es5",
     "resolveJsonModule": true,
     "esModuleInterop": true,
+      // Ensure that .d.ts files are created by tsc, but not .js files
+    "declaration": true,
+    "emitDeclarationOnly": true,
+    // Ensure that Babel can safely transpile files in the TypeScript project
+    "isolatedModules": true
   },
   "include": ["./src/**/*"]
 }
